[feat] Initial loading state for cold-start
